refactor: Rename client to gateway
